+++ src/wordpress.php
@@ -106,278 +106,36 @@ function get_allowed_use_subdirectories(): array {
 }
 
 /**
- * Ensures WordPress file are correctly unzipped and placed.
- *
- * If a different version of WordPress is already installed, then
- * it will be removed.
- *
- * @param string|null $version    The WordPress version to set up
- *                                the files for.
- *
- * @return bool Always `true` to indicate files are in place.
- */
-function ensure_wordpress_files( $version = null ): bool {
-	// By default, download the latest WordPress version.
-	$source_url = 'https://wordpress.org/latest.zip';
-
-	if ( $version !== null ) {
-		// The provided WordPress version will override any env defined version.
-		$source_url = "https://wordpress.org/wordpress-$version.zip";
-	} else {
-		// If set, then use the WordPress version defined by the env.
-		$env_wp_version = getenv( 'SLIC_WP_VERSION' );
-		$version        = $env_wp_version;
-		if ( ! empty( $env_wp_version ) ) {
-			$source_url = "https://wordpress.org/wordpress-$env_wp_version.zip";
-		}
-	}
-
-	$version = $version ?: 'latest';
-
-	debug( "Checking if WordPress version $version is installed and configured ... " . PHP_EOL );
-
-	if ( $version === 'latest' ) {
-		debug( "Resolving latest version to a semantic version string ..." . PHP_EOL );
-		$version = get_wordpress_latest_version();
-	}
-
-	$wp_root_dir  = getenv( 'SLIC_WP_DIR' );
-	$version_file = $wp_root_dir . '/wp-includes/version.php';
-
-	// Check only if the specified version is not latest.
-	if ( is_file( $version_file ) ) {
-		include_once $version_file;
-
-		// `$wp_version` is globally defined in the `wp-includes/version.php` file.
-		if ( isset( $wp_version ) && version_compare( $wp_version, $version ) === 0 ) {
-
-			debug( "WordPress current version ($wp_version) matches the requested one ($version)" . PHP_EOL );
-
-			return true;
-		}
-
-		if ( isset( $wp_version ) ) {
-			echo "WordPress current version ($wp_version) does not match the requested one ($version), removing WordPress directory ... ";
-		}
-
-		// Remove the previous version of WordPress.
-		quietly_tear_down_stack();
-
-		if ( ! rrmdir( $wp_root_dir ) ) {
-			magenta( "Failed to remove the previous WordPress directory, try manually." . PHP_EOL );
-			exit( 1 );
-		}
-
-		echo light_cyan( "done" . PHP_EOL );
-	} else {
-		debug( "Previous WordPress directory not found." . PHP_EOL );
-	}
-
-	// Tear down the stack to avoid containers from locking the bound directories.
-	quietly_tear_down_stack();
-
-	// Ensure the destination directory exists.
-	if ( ! is_dir( $wp_root_dir ) && ! mkdir( $wp_root_dir, 0755, false ) && ! is_dir( $wp_root_dir ) ) {
-		echo magenta( "Failed to create WordPress root directory {$wp_root_dir}" );
-		exit( 1 );
-	}
-
-	// Download WordPress.
-	$zip_file = cache( "/wordpress/wordpress-$version.zip" );
-	if ( ! is_file( $zip_file ) ) {
-		debug( "WordPress zip file $zip_file not found." . PHP_EOL );
-
-		$zip_file = download_file( $source_url, $zip_file );
-
-		if ( $zip_file === false ) {
-			echo magenta( "Failed to download WordPress file from $source_url." );
-			exit( 1 );
-		}
-	}
-
-	// Unzip WordPress.
-	if ( ! is_file( $wp_root_dir . '/wp-load.php' ) && ! unzip_file( $zip_file, $wp_root_dir ) ) {
-		echo magenta( "Failed to extract WordPress file $zip_file to $wp_root_dir." );
-		exit( 1 );
-	}
-
-	return true;
-}
-
-/**
- * Ensures WordPress is correctly configured.
- *
- * @return bool Always `true` to indicate WordPress
- *              is set up correctly.
- */
-function ensure_wordpress_configured(): bool {
-	$wp_root_dir    = getenv( 'SLIC_WP_DIR' );
-	$wp_config_file = $wp_root_dir . '/wp-config.php';
-
-	if ( is_file( $wp_config_file ) ) {
-		debug( "Found $wp_config_file, assuming WordPress already configured." . PHP_EOL );
-
-		// If the wp-config.php file already exists, assume WordPress is already configured correctly.
-		return true;
-	}
-
-	$wp_config_sample_file = $wp_root_dir . '/wp-config-sample.php';
-
-	if ( ! is_file( $wp_config_sample_file ) ) {
-		echo magenta( "Config sample file $wp_config_sample_file not found." );
-		exit( 1 );
-	}
-
-	$wp_config_contents = file_get_contents( $wp_config_sample_file );
-
-	if ( empty( $wp_config_contents ) ) {
-		echo magenta( "Config sample file $wp_config_sample_file could not be read or is empty." );
-		exit( 1 );
-	}
-
-	debug( "Setting up credentials in $wp_config_file ..." . PHP_EOL );
-
-	// Set up the db credentials, rely on the placeholders that come with a default WordPress installation.
-	$wp_config_contents = str_replace( [
-		'<?php',
-		"'database_name_here'",
-		"'username_here'",
-		"'password_here'",
-		"'localhost'"
-	], [
-		"<?php\n\nfunction slic_env( \$key, \$default ){\n\treturn getenv( \$key ) ?: \$default;\n}\n",
-		"'" . get_db_name() . "'",
-		"'" . get_db_user() . "'",
-		"'" . get_db_password() . "'",
-		"slic_env( 'DB_HOST', 'db' )"
-	], $wp_config_contents );
-
-	debug( "Setting up salts in $wp_config_file ..." . PHP_EOL );
-
-	// As is common practice, use the "That's all, stop editing! Happy publishing" line as a marker.
-	$marker = "/* That's all, stop editing! Happy publishing. */";
-
-	if ( strpos( $wp_config_contents, $marker ) === false ) {
-		echo magenta( "Config sample file $wp_config_sample_file does not contain marker line." );
-		exit( 1 );
-	}
-
-	// Generate salts, there are 8 of them, each should be different.
-	for ( $i = 0; $i < 8; $i ++ ) {
-		// Cryptographically weak, but fine in testing environment.
-		$salt = hash( 'sha256', (string) microtime( true ) . mt_rand( 1, PHP_INT_MAX ) );
-		// Use `preg_replace` to limit the replacements.
-		$wp_config_contents = preg_replace( '/put your unique phrase here/', $salt, $wp_config_contents, 1 );
-	}
-
-	debug( "Setting up config extras in $wp_config_file ..." . PHP_EOL );
-
-	$config_extras      = <<< CONFIG_EXTRAS
-\$scheme = empty( \$_SERVER['HTTPS'] ) ? 'http' : 'https';
-\$url    = isset( \$_SERVER['HTTP_HOST'] ) ? \$_SERVER['HTTP_HOST'] : 'wordpress.test';
-define( 'WP_HOME', \$scheme . '://' . \$url );
-define( 'WP_SITEURL', \$scheme . '://' . \$url );
-define( 'WP_REDIS_HOST', 'redis' );
-define( 'WP_REDIS_PORT', 6379 );
-define( 'TRIBE_NO_FREEMIUS', true );
-define( 'WP_DEBUG_DISPLAY', true );
-define( 'WP_DEBUG_LOG', true );
-CONFIG_EXTRAS;
-	$wp_config_contents = str_replace( $marker, $marker . PHP_EOL . $config_extras, $wp_config_contents );
-
-	if ( ! file_put_contents( $wp_config_file, $wp_config_contents, LOCK_EX ) ) {
-		echo magenta( "Failed to write $wp_config_file file." );
-		exit( 1 );
-	}
-
-	debug( "WordPress $wp_config_file updated." . PHP_EOL );
-
-	return true;
-}
-
-/**
  * Ensures WordPress is correctly installed.
  *
  * @return bool Always `true` to indicate WordPress is
  *              correctly installed.
  */
 function ensure_wordpress_installed(): bool {
-	setup_slic_env( root() );
+	if ( slic_realtime()( cli_command( [ 'core', 'is-installed' ] ) ) === 0 ) {
+		debug( "WordPress is already installed." . PHP_EOL );
 
-	// Bring up the database.
-	ensure_db_service_ready();
+		return true;
+	}
 
-	debug( "Trying to connect to the database ..." . PHP_EOL );
-
-	// Run a query to check for the installation.
-	$db = get_localhost_db_handle();
-
-	if ( ! $db instanceof \mysqli ) {
-		echo magenta( 'Failed to connect to WordPress database: ' . mysqli_connect_error() );
+	$install = [
+		'core',
+		'install',
+		'--url=http://wordpress.test',
+		'--title=Slic',
+		'--admin_user=admin',
+		'--admin_password=password',
+		'--admin_email=admin@wordpress.test',
+		'--skip-email',
+	];
+	if ( slic_realtime()( cli_command( $install ) ) !== 0 ) {
+		// There will be debug detailing the issue.
+		echo magenta( "Failed to install WordPress." );
 		exit( 1 );
 	}
 
-	debug( "Getting tables list ..." . PHP_EOL );
-
-	// Check if the default tables are there or not.
-	$tables = $db->query( 'SHOW TABLES' );
-
-	if ( ! $tables instanceof \mysqli_result ) {
-		echo magenta( 'Failed to query the WordPress database: ' . mysqli_error( $db ) );
-		exit( 1 );
-	}
-
-	$tables_list = $tables->fetch_all( MYSQLI_NUM );
-	if ( ! empty( $tables_list ) ) {
-		$default_tables = get_default_tables_list();
-		$tables_list    = array_column( $tables_list, 0 );
-
-		if ( count( array_diff( $default_tables, $tables_list ) ) === 0 ) {
-
-			debug( "Default tables found: assuming WordPress is installed." . PHP_EOL );
-
-			return true;
-		}
-	}
-
-	debug( "Default tables not found: assuming WordPress is not installed." . PHP_EOL );
-
-	$wp_root_dir  = getenv( 'SLIC_WP_DIR' );
-	$install_file = realpath( $wp_root_dir . '/wp-admin/install.php' );
-
-	if ( ! is_file( $install_file ) ) {
-		echo magenta( "WordPress installation file $install_file not found." );
-		exit( 1 );
-	}
-
-	// In a separate process, call the installation file directly setting up the expected request vars.
-	$code = 'putenv( "DB_HOST=' . get_localhost_db_host() . '" ); ' .
-	        '$_GET["step"] = 2; ' .
-	        '$_POST["weblog_title"] = "Slic Test Site"; ' .
-	        '$_POST["user_name"] = "admin"; ' .
-	        '$_POST["admin_password"] = "password"; ' .
-	        '$_POST["admin_password2"] = "password"; ' .
-	        '$_POST["admin_email"] = "admin@wordpress.test"; ' .
-	        '$_POST["blog_public"] = 1; ' .
-	        'function wp_mail(){ return true; } ' . // It's pluggable, this will mute it.
-	        'include "' . $install_file . '";';
-
-	$command = escapeshellarg( PHP_BINARY ) . ' -r \'' . $code . '\'';
-
-	debug( "Installing WordPress with command $command ... " . PHP_EOL );
-
-	exec( $command, $output, $status );
-
-	$error_lines       = array_filter( $output, static function ( $line ) {
-		return strpos( $line, 'Error' ) !== false;
-	} );
-	$output_has_errors = count( $error_lines );
-
-	if ( $status !== 0 || $output_has_errors ) {
-		$circa_error_lines = array_map( static function ( $line_number, $line ) use ( $output ) {
-			return strip_tags( implode( PHP_EOL, array_slice( $output, $line_number, 10 ) ) );
-		}, array_keys( $error_lines ), $error_lines );
-		echo magenta( "WordPress installation failed with message(s):" . PHP_EOL . implode( PHP_EOL, $circa_error_lines ) );
+	if ( slic_realtime()( cli_command( [ 'core', 'is-installed' ] ) ) !== 0 ) {
+		echo magenta( "Failed to check WordPress is installed." );
 		exit( 1 );
 	}
 
@@ -448,8 +206,7 @@ function get_wordpress_latest_version(): string {
  * @return bool Always `true` to indicate success.
  */
 function ensure_wordpress_ready( string $version = null ): bool {
-	ensure_wordpress_files( $version );
-	ensure_wordpress_configured();
+	ensure_services_running( [ 'slic', 'wordpress' ] );
 	ensure_wordpress_installed();
 
 	return true;