+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - TBD
+
+* Change - Leverage Docker Compose syntax to manage network aliases and service dependencies. Clean up a remove code.
+* Add the `dc` command to run Docker Compose commands in the stack.
+
 ## [1.2.2] - 2023-02-24
 
 * Change - No longer build gd in the slick docker container with AVIF, fixing external dependency on the currently down code.videolan.org, and speeding up builds - [more info](https://github.com/mlocati/docker-php-extension-installer#configuration).